reporting/bufferSize.html

<!DOCTYPE HTML>
<meta charset=utf-8>
<title>Reporting: Buffer size</title>
<link rel="author" title="Paul Meyer" href="paulmeyer@chromium.org">
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<script src="/resources/testdriver.js"></script>
<script>
  async_test(async function(test) {
    for (i = 0; i != 110; ++i)
      await test_driver.generate_test_report("" + i);

    var observer = new ReportingObserver(function(reports) {
      test.step(function() {
        assert_equals(reports.length, 100);
        for(i = 0; i != 100; ++i) {
          assert_equals(reports[i].body.message, "" + (i + 10));
        }
      });

      test.done();
    }, {buffered: true});
    observer.observe();
  }, "Buffer size");
</script>
